Revert "Refactor:extract common module postgresql-common (#3725)"
This reverts commit d9fcda13257808f5e40e4565d7240ba669c0ac3c.
